Added and applied linting
The "code quality" check was failing in the pipeline because it was
enforcing some linting rules that were rot in the repo.

I created an .eslintrc.js file with the few rules I know the pipeline
was trying to enforce.

I applied this to the whole project `eslint . --fix`, so it touched
a lot of files.