Add support for querying the Tenor API
Updated the @jych/gif-provider dependency to @0.5.0, which adds the
TenorGifProvider.

Since we now want to query multiple GifProviders at once, I wrapped the
Gihpy and Tenor providers in a CompositeGifProvider. The interface
remains the same so no other code changes needed.

The API key used for the Tenor API is unique to this project as I
registered a new one.